+++ magda-storage-api/src/test/storage.spec.ts
@@ -59,6 +59,7 @@ describe("Storage API tests", () => {
     const registryApiUrl = "http://registry.example.com";
     const jwtSecret = "squirrel";
     let registryScope: nock.Scope;
+    const uploadLimit = "100mb";
 
     before(() => {
         minioClient.makeBucket(bucketName, (err: Error) => {
@@ -80,7 +81,8 @@ describe("Storage API tests", () => {
                 authApiUrl,
                 registryApiUrl,
                 jwtSecret,
-                tenantId: 0
+                tenantId: 0,
+                uploadLimit
             })
         );
     });
@@ -188,6 +190,62 @@ describe("Storage API tests", () => {
                         .query({ recordId: "storage-test-dataset-id-success" })
                         .send("LALALALALALALALALA")
                         .expect(401, "Not authorized.")
+                );
+            });
+        });
+
+        describe("Upload from the browser", () => {
+            describe("Upload some files", () => {
+                it("Not as an admin", () => {
+                    return mockAuthorization(
+                        authApiUrl,
+                        false,
+                        jwtSecret,
+                        request(app)
+                            .post("/v0/upload/" + bucketName)
+                            .field(
+                                "originalname",
+                                "test-browser-upload-no-admin"
+                            )
+                            .attach("image", "src/test/test_image.jpg")
+                            .expect(401, "Not authorized.")
+                    );
+                });
+
+                it("As an admin", () => {
+                    const bananadance: Buffer = fs.readFileSync(
+                        "src/test/bananadance.gif"
+                    );
+                    return mockAuthorization(
+                        authApiUrl,
+                        true,
+                        jwtSecret,
+                        request(app)
+                            .post("/v0/upload/" + bucketName)
+                            .attach("text", "src/test/test_csv_1.csv")
+                            .accept("csv")
+                            .attach("image", bananadance, "bananadance.gif")
+                            .accept("gif")
+                            .expect(200)
+                    ).then(_res => {
+                        return request(app)
+                            .get("/v0/" + bucketName + "/bananadance.gif")
+                            .accept("gif")
+                            .expect(200)
+                            .expect(bananadance);
+                    });
+                });
+            });
+
+            it("Upload no files", () => {
+                return mockAuthorization(
+                    authApiUrl,
+                    true,
+                    jwtSecret,
+                    request(app)
+                        .post("/v0/upload/" + bucketName)
+                        .field("originalname", "test-browser-upload-1")
+                        .expect(400)
                 );
             });
         });
@@ -335,6 +393,54 @@ describe("Storage API tests", () => {
                         .expect(img)
                         .expect("Content-Type", "image/jpg")
                 );
+            });
+        });
+
+        describe("Binary content", () => {
+            it("JPG Image", () => {
+                const img: Buffer = fs.readFileSync("src/test/test_image.jpg");
+                return mockAuthorization(
+                    authApiUrl,
+                    true,
+                    jwtSecret,
+                    request(app)
+                        .put("/v0/" + bucketName + "/binary-content-jpg")
+                        .set("Accept", "image/jpg")
+                        .set("Content-Type", "image/jpg")
+                        .send(img)
+                        .expect(200)
+                ).then(_res => {
+                    return request(app)
+                        .get("/v0/" + bucketName + "/binary-content-jpg")
+                        .set("Accept", "image/jpg")
+                        .expect(200)
+                        .expect(img)
+                        .expect("Content-Type", "image/jpg");
+                });
+            });
+
+            it("Bananadance GIF", () => {
+                const bananadance: Buffer = fs.readFileSync(
+                    "src/test/bananadance.gif"
+                );
+                return mockAuthorization(
+                    authApiUrl,
+                    true,
+                    jwtSecret,
+                    request(app)
+                        .put("/v0/" + bucketName + "/binary-content-gif")
+                        .set("Accept", "image/gif")
+                        .set("Content-Type", "image/gif")
+                        .send(bananadance)
+                        .expect(200)
+                ).then(_res => {
+                    return request(app)
+                        .get("/v0/" + bucketName + "/binary-content-gif")
+                        .set("Accept", "image/gif")
+                        .expect(200)
+                        .expect(bananadance)
+                        .expect("Content-Type", "image/gif");
+                });
             });
         });