+++ CHANGES.md
@@ -23,6 +23,7 @@ UI:
 -   Make the global notification banner configurable
 -   Removed all references to the DTA Design System react components, so that all styles come through our SCSS compilation, which should make SCSS smaller and more consistent.
 -   Swap the order of custodian & team dropdown on add dataset page
+-   Updated Add Dataset Welcome screen options UI design
 
 Storage:
 
@@ -33,6 +34,11 @@ Storage:
 -   Add endpoint to create a bucket
 -   Restrict file upload to admins only
 -   Add authorization to GET endpoint
+-   Support multipart upload
+
+Gateway:
+
+-   Add /data to ckan URL, remove the `came_from` param
 
 Gateway: